requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
pillow
tqdm
matplotlib
fastapi
uvicorn
rembg
python-multipart
